backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
scikit-learn==1.5.2
sentence-transformers==3.0.1
rapidfuzz==3.9.5
python-docx==1.1.2
PyPDF2==3.0.1
numpy==1.26.4
spacy==3.7.5
pytest==8.2.1
python-dotenv==1.0.1
python-multipart==0.0.9

